Move SimCardDemo > CompositeView; update directory
